fixed link redirects
they were pointing to "new" urls (created as part of the tumblr -> jekyll port) that I changed back to the old URLs because I couldn't figure out how to do the redirects on the permalinks. Now the posts have the old tumblr-style URLs and the links to those old posts in the body of other posts point to the old tumblr-style URLs.